+++ .idea/libraries/mysql_connector_java_8_0_24.xml
@@ -4,8 +4,6 @@
       <root url="jar://$USER_HOME$/Desktop/mysql-connector-java-8.0.24/mysql-connector-java-8.0.24.jar!/" />
     </CLASSES>
     <JAVADOC />
-    <SOURCES>
-      <root url="jar://$USER_HOME$/.ideaLibSources/mysql-connector-java-8.0.24-sources.jar!/" />
-    </SOURCES>
+    <SOURCES />
   </library>
 </component>
