+++ .idea/sqldialects.xml
@@ -1,8 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="SqlDialectMappings">
-    <file url="file://$PROJECT_DIR$/src/main/java/db/migration/V1_1__create_db.sql" dialect="GenericSQL" />
-    <file url="file://$PROJECT_DIR$/src/main/java/db/migration/V1_2__create_table_user.sql" dialect="GenericSQL" />
+    <file url="file://$PROJECT_DIR$/src/main/resources/db/migration/V1_1__create_table_user.sql" dialect="GenericSQL" />
     <file url="PROJECT" dialect="MySQL" />
   </component>
 </project>
