+++ src/main/docker-compose.yml
@@ -4,7 +4,7 @@ services:
     image: mysql:5.7
     restart: always
     environment:
-      MYSQL_DATABASE: 'filekeeper'
+      MYSQL_DATABASE: 'filekeeper2'
       # So you don't have to use root, but you can if you like
       MYSQL_USER: 'root'
       # You can use whatever password you like
