+++ src/main/java/Main.java
@@ -1,12 +1,9 @@
-package main.java;
-
-import сlient.Client;
-import server.Server;
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Parent;
 import javafx.scene.Scene;
 import javafx.stage.Stage;
+import org.flywaydb.core.Flyway;
 
 
 public class Main extends Application {
@@ -14,12 +11,19 @@ public class Main extends Application {
     public void start(Stage primaryStage) throws Exception {
         //Server.startServer();
 
+
         Parent root = FXMLLoader.load(getClass().getResource("windows/main.fxml"));
         primaryStage.setTitle("FileKeeper");
         primaryStage.setScene(new Scene(root, 650, 400));
         primaryStage.setResizable(false);
         primaryStage.show();
-        primaryStage.setOnCloseRequest(event -> Client.closeConnection());
+        migrateDatabase();
 
+    }
+
+    public void migrateDatabase() {
+        Flyway flyway = new Flyway();
+        flyway.setDataSource("jdbc:mysql://localhost:3306/test", "root", "password");
+        flyway.migrate();
     }
 }
