--- src/main/java/db/migration/V1_1__create_db.sql
@@ -1,1 +1,0 @@
-CREATE SCHEMA `filekeeper`;
