+++ src/main/java/server/Sceneable.java
@@ -1,0 +1,7 @@
+package server;
+
+import java.awt.*;
+
+public interface Sceneable {
+    void openNewScene(String window);
+}
