+++ src/main/resources/db/migration/V1_1__create_table_user.sql
@@ -1,0 +1,7 @@
+CREATE TABLE user IF NOT EXISTS
+(
+    iduser INT NOT NULL AUTO_INCREMENT,
+    user_name VARCHAR(45) NOT NULL,
+    user_pass VARCHAR(45) NOT NULL,
+    PRIMARY KEY (iduser)
+);
